# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: PinIt Security & Antivirus Detection API
  description: |
    Enterprise-grade security analysis API for scanning uploaded files and URLs for malware,
    viruses, phishing, macros, obfuscated scripts, and suspicious behavior.
  version: 1.0.0
  contact:
    name: Security Engineering Team
    url: https://pinit.security
servers:
  - url: http://localhost:5000/api/v1
    description: Local Development Server
paths:
...
  /auth/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Authenticated.
  /auth/me:
    get:
      summary: Get current authenticated user profile
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        '200':
          description: Profile retrieved.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...